requirements.txt
Django==2.0.1
djangorestframework==3.7.7
python-dateutil==2.6.1
virtualenv==15.1.0